Simplify test rpc crate
