+++ crates/private/support/sel4-simple-task/rpc/Cargo.toml
@@ -16,10 +16,7 @@ authors = ["Nick Spinale <nick.spinale@coliasgroup.com>"]
 edition = "2024"
 license = "BSD-2-Clause"
 
-[features]
-postcard = ["dep:serde", "dep:postcard"]
-
 [dependencies]
-postcard = { version = "1.1.3", default-features = false, optional = true }
+postcard = { version = "1.1.3", default-features = false }
 sel4 = { path = "../../../../sel4" }
-serde = { version = "1.0.228", default-features = false, optional = true }
+serde = { version = "1.0.228", default-features = false }
